app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "stats bot"
authors = ["Flict-dev <big.bob@gmail.com>", "Ramnck <aboba@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.19.2"
aiogram = "3.0.0b6"
python-decouple = "^3.6"
asyncpg = "^0.27.0"
pydantic = "^1.10.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.11.4"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
